requirements.txt
aiobotocore==2.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
backoff==1.11.1
billiard==4.2.1
boto3==1.38.43
botocore==1.38.43
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
cellmap-schemas==0.8.1
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-files==5.4.0
cloud_volume==12.3.1
cloudpickle==3.1.1
colorama==0.4.6
compressed_segmentation==2.3.2
configobj==5.0.9
contourpy==1.3.2
crc32c==2.7.1
cryptography==45.0.4
cycler==0.12.1
dask==2025.5.1
deflate==0.8.0
Deprecated==1.2.18
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distributed==2025.5.1
distro==1.9.0
dm3_lib @ git+https://github.com/piraynal/pyDM3reader.git@3a56ac751a693b7130ae9d9360a0fee655092cbf
dpath==2.2.0
DracoPy==1.5.0
dulwich==0.23.0
dvc==3.60.1
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
eval-type-backport==0.1.3
fasteners==0.19
fastremap==1.17.1
fibsem-tools==7.0.5
filelock==3.18.0
flatten-dict==0.4.2
flexcache==0.3
flexparser==0.4
flufl.lock==8.2.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
funcy==2.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
greenlet==3.2.3
gto==1.7.2
hydra-core==1.3.2
idna==3.10
imagecodecs==2025.3.30
inflection==0.5.1
iniconfig==2.1.0
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
json5==0.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
locket==1.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
microviewer==1.16.0
mrcfile==1.5.4
msgpack==1.1.1
multidict==6.5.0
multiprocess==0.70.18
networkx==3.5
numcodecs==0.15.1
numpy==2.3.1
omegaconf==2.3.0
orjson==3.10.18
osteoid==0.3.2
packaging==25.0
pandas==2.3.0
partd==1.4.2
pathos==0.3.4
pathspec==0.12.1
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
pluggy==1.6.0
posix_ipc==1.2.0
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ome-ngff==0.6.0
pydantic-zarr==0.7.0
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
pysimdjson==7.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-jsonschema-objects==0.5.7
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3fs==2025.5.1
s3transfer==0.13.0
scipy==1.16.0
scmrepo==3.3.11
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simplejpeg==1.8.2
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
sqltrie==0.11.2
tabulate==0.9.0
tblib==3.1.0
tenacity==9.1.2
tifffile==2025.6.11
tomlkit==0.13.3
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
wrapt==1.17.2
xarray==2025.6.1
xarray-multiscale==2.1.0
xarray-ome-ngff==3.1.1
yarl==1.20.1
zarr==2.18.7
zc.lockfile==3.0.post1
zict==3.0.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
